cybrhunter/requirements.txt
pip>=9.0.3
kafka>=1.3.5
pika>=0.12.0
pyyaml>=3.13